requirements.txt
Django==4.2.16
djangorestframework==3.15.2
redis==5.2.0
pyjwt==2.9.0
django-environ==0.11.2
celery==5.4.0
pika==1.3.2
requests==2.32.3
pybreaker==1.2.0
drf-yasg==1.21.8
gunicorn==23.0.0
cryptography==43.0.3